+++ InstagramCarouselGenerator.lrplugin/Info.lua
@@ -12,8 +12,8 @@ return {
     LrSdkVersion = 15.1,
     LrSdkMinimumVersion = 11.0,
     
-    -- Plugin Information
-    LrPluginName = "Instagram Carousel Generator",
+    -- Plugin Information (localized with ZStrings)
+    LrPluginName = LOC "$$$/InstagramCarousel/PluginName=Instagram Carousel Generator",
     LrPluginInfoUrl = "https://github.com/manuzzi/LightroomClassicCarouselPlugin",
     
     -- Unique Plugin Identifier (reverse domain notation)
@@ -25,21 +25,21 @@ return {
     -- Plugin Version
     VERSION = { 
         major = 1, 
-        minor = 3, 
+        minor = 4, 
         revision = 0,
         build = 0,
     },
     
-    -- Export Service Provider
+    -- Export Service Provider (localized)
     LrExportServiceProvider = {
-        title = "Instagram Carousel",
+        title = LOC "$$$/InstagramCarousel/ExportServiceName=Instagram Carousel",
         file = "InstagramCarouselExportServiceProvider.lua",
     },
     
-    -- Help Menu Items
+    -- Help Menu Items (localized)
     LrHelpMenuItems = {
         {
-            title = "About Instagram Carousel Generator",
+            title = LOC "$$$/InstagramCarousel/Help/About=About Instagram Carousel Generator",
             file = "Documentation/About.txt",
         },
     },